src/main/java/javaguru/sunday/student_deniss_lobacs/lesson_14/level_1_and_2/FruitStorage.java
package javaguru.sunday.student_deniss_lobacs.lesson_14.level_1_and_2;

import java.util.ArrayList;
import java.util.List;

public class FruitStorage {

    public List<Apple> getAllApples() {
        List<Apple> apples = new ArrayList<>();
        apples.add(new Apple("red", 100));
        apples.add(new Apple("red", 160));
        apples.add(new Apple("red", 200));
        apples.add(new Apple("green", 200));
        apples.add(new Apple("green", 90));
        apples.add(new Apple("green", 50));
        apples.add(new Apple("yellow", 50));
        apples.add(new Apple("yellow", 170));
        return apples;
    }

    public List<Apple> findAllGreenApples(List<Apple> apples){
        List<Apple> greenApples = new ArrayList<>();
        for (Apple apple : apples) {
            if(apple.getColor().equals("green")){
                greenApples.add(apple);
            }
        }
        return greenApples;
    }

    public List<Apple> findAllRedApples(List<Apple> apples){
        List<Apple> redApples = new ArrayList<>();
        for (Apple apple : apples) {
            if(apple.getColor().equals("red")){
                redApples.add(apple);
            }
        }
        return redApples;
    }

    public List<Apple> findApplesByColor(List<Apple> inventory, String color) {
        List<Apple> result = new ArrayList<>();
        for(Apple apple : inventory) {
            if(apple.getColor().equals(color)) {
                result.add(apple);
            }
        }
        return result;
    }

    public List<Apple> findApplesByWeight(List<Apple> inventory, int weight) {
        List<Apple> result = new ArrayList<>();
        for(Apple apple : inventory) {
            if(apple.getWeight() > weight) {
                result.add(apple);
            }
        }
        return result;
    }

    public List<Apple> findApples(List<Apple> inventory, ApplePredicate p) {
        List<Apple> result = new ArrayList<>();
        for (Apple apple : inventory) {
            if (p.test(apple)) {
                result.add(apple);
            }
        }
        return result;
    }

}
